centralize AtlasDbConfig loading
