+++ atlasdb-cli/build.gradle
@@ -11,9 +11,6 @@ dependencies {
   compile project(':atlasdb-impl-shared')
   compile 'io.airlift:airline:0.7'
   compile 'com.google.dagger:dagger:2.0.2'
-  compile 'com.fasterxml.jackson.core:jackson-databind:' + libVersions.jackson
-  compile 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:' + libVersions.jackson
-  compile 'io.dropwizard:dropwizard-jackson:' + libVersions.dropwizard
 
   testCompile project(':atlasdb-rocksdb')
