+++ atlasdb-cli/src/main/java/com/palantir/atlasdb/cli/services/ServicesConfigModule.java
@@ -17,19 +17,14 @@ package com.palantir.atlasdb.cli.services;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.Iterator;
 
 import javax.inject.Singleton;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
-import com.google.common.base.Strings;
 import com.palantir.atlasdb.config.AtlasDbConfig;
+import com.palantir.atlasdb.config.AtlasDbConfigs;
 
 import dagger.Module;
 import dagger.Provides;
-import io.dropwizard.jackson.DiscoverableSubtypeResolver;
 
 @Module
 public class ServicesConfigModule {
@@ -37,39 +32,7 @@ public class ServicesConfigModule {
     private final ServicesConfig config;
 
     public static ServicesConfigModule create(File configFile, String configRoot) throws IOException {
-        ObjectMapper configMapper = new ObjectMapper(new YAMLFactory());
-        configMapper.setSubtypeResolver(new DiscoverableSubtypeResolver());
-        JsonNode node = getConfigNode(configMapper, configFile, configRoot);
-        AtlasDbConfig config = configMapper.treeToValue(node, AtlasDbConfig.class);
-        return ServicesConfigModule.create(config);
-    }
-
-    private static JsonNode getConfigNode(ObjectMapper configMapper, File configFile, String configRoot) throws IOException {
-        JsonNode node = configMapper.readTree(configFile);
-        if (Strings.isNullOrEmpty(configRoot)) {
-            return node;
-        } else {
-            JsonNode rootNode = findRoot(node, configRoot);
-            if (rootNode != null) {
-                return rootNode;
-            }
-            throw new IllegalArgumentException("Could not find " + configRoot + " in yaml file " + configFile);
-        }
-    }
-
-    private static JsonNode findRoot(JsonNode node, String configRoot) {
-        if (node.has(configRoot)) {
-            return node.get(configRoot);
-        } else {
-            Iterator<String> iter = node.fieldNames();
-            while (iter.hasNext()) {
-                JsonNode root = findRoot(node.get(iter.next()), configRoot);
-                if (root != null) {
-                    return root;
-                }
-            }
-            return null;
-        }
+        return ServicesConfigModule.create(AtlasDbConfigs.load(configFile, configRoot));
     }
 
     public static ServicesConfigModule create(AtlasDbConfig atlasDbConfig) {
