+++ atlasdb-console/build.gradle
@@ -7,9 +7,6 @@ dependencies {
     compile('commons-cli:commons-cli:1.1')
     compile('jline:jline:2.12')
     compile('org.codehaus.groovy:groovy-all:' + libVersions.groovy)
-    compile 'com.fasterxml.jackson.core:jackson-databind:' + libVersions.jackson
-    compile 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:' + libVersions.jackson
-    compile 'io.dropwizard:dropwizard-jackson:' + libVersions.dropwizard
 
     testCompile('org.jmock:jmock-legacy:' + libVersions.jmock) {
       exclude(module: 'jmock')
