+++ atlasdb-server/build.gradle
@@ -6,3 +6,10 @@ dependencies {
     compile project(':atlasdb-config')
     compile 'javax.inject:javax.inject:1'
 }
+
+configurations.matching({ it.name in ['compile', 'runtime'] }).all {
+    resolutionStrategy {
+        force 'com.google.code.findbugs:jsr305:' + libVersions.jsr305
+        force 'org.slf4j:slf4j-api:' + libVersions.slf4j
+    }
+}
